Refactor jestRunner import
