+++ package.json
@@ -14,7 +14,6 @@
     "lint-ts": "npm run format-check && tslint --project packages/unmock-types/ --config tslint.json && tslint --project packages/openapi-refinements/ --config tslint.json && tslint --project packages/unmock/tsconfig.node.json --config tslint.json && tslint --project packages/unmock/tsconfig.browser.json --config tslint.json && tslint --project packages/unmock-cli/ --config tslint.json && tslint --project packages/unmock-core/ --config tslint.json && tslint --project packages/unmock-jest/ --config tslint.json && tslint --project packages/unmock-node/ --config tslint.json && tslint --project packages/unmock-server/ --config tslint.json && tslint --project packages/unmock-browser/ --config tslint.json",
     "lint-ts-fix": "npm run format && tslint --fix --project packages/unmock-types/ --config tslint.json && tslint --fix --project packages/openapi-refinements/ --config tslint.json && tslint --fix --project packages/unmock/tsconfig.node.json --config tslint.json && tslint --fix --project packages/unmock/tsconfig.browser.json --config tslint.json && tslint --fix --project packages/unmock-cli/ --config tslint.json && tslint --fix --project packages/unmock-core/ --config tslint.json && tslint --fix --project packages/unmock-jest/ --config tslint.json && tslint --fix --project packages/unmock-node/ --config tslint.json && tslint --fix --project packages/unmock-server/ --config tslint.json && tslint --fix --project packages/unmock-browser/ --config tslint.json",
     "test": "jest --verbose",
-    "test:mocha": "mocha -r ts-node/register packages/**/*.mocha-only.ts",
     "test:debug": "DEBUG=unmock* jest --verbose",
     "test:clean": "jest --clearCache",
     "test:watch": "jest --verbose --watchAll",
