+++ packages/unmock/src/__tests__/end-to-end/readme/introductory-example.test.js
@@ -1,7 +1,7 @@
 const unmock = require("unmock");
 const { nock, transform, u } = unmock;
 const { withCodes } = transform;
-const runner = require("unmock-runner").default;
+const jestRunner = require("../../../../../unmock-runner/src/jestRunner").default;
 
 nock("https://zodiac.com", "zodiac")
   .get("/horoscope/{sign}")
@@ -23,18 +23,6 @@ beforeAll(() => {
   zodiac = unmock.default.on().services.zodiac;
 });
 afterAll(() => unmock.default.off());
-
-const jestRunner = (fn, options) => async cb => {
-  return runner(e => e.constructor.name === "JestAssertionError")(
-    unmock.default,
-  )(meeshkanCallback => {
-    const asJestCallback = () => {
-      meeshkanCallback.success();
-    };
-    asJestCallback.fail = meeshkanCallback.fail;
-    return fn ? fn(asJestCallback) : undefined;
-  }, options)(cb ? { success: cb, fail: cb.fail } : undefined);
-};
 
 describe("getHoroscope", () => {
   it(
