+++ packages/unmock/src/__tests__/end-to-end/readme/specifying-the-reply.test.js
@@ -1,7 +1,7 @@
 const unmock = require("unmock");
 const { nock, transform, u } = unmock;
 const { withCodes } = transform;
-const runner = require("unmock-runner").default;
+const jestRunner = require("../../../../../unmock-runner/src/jestRunner").default;
 
 unmock
   .nock("http://www.foo.com")
@@ -20,18 +20,6 @@ beforeAll(() => {
   zodiac = unmock.default.on();
 });
 afterAll(() => unmock.default.off());
-
-const jestRunner = (fn, options) => async cb => {
-  return runner(e => e.constructor.name === "JestAssertionError")(
-    unmock.default,
-  )(meeshkanCallback => {
-    const asJestCallback = () => {
-      meeshkanCallback.success();
-    };
-    asJestCallback.fail = meeshkanCallback.fail;
-    return fn ? fn(asJestCallback) : undefined;
-  }, options)(cb ? { success: cb, fail: cb.fail } : undefined);
-};
 
 describe("getName", () => {
   it(
